CAppSettings: change dZoomX and dZoomY from type float to double, as these are used as such
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@5305 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/mpc-hc/AppSettings.h
@@ -384,8 +384,8 @@ public:
     bool            fRememberWindowPos;
     bool            fRememberWindowSize;
     bool            fSavePnSZoom;
-    float           dZoomX;
-    float           dZoomY;
+    double          dZoomX;
+    double          dZoomY;
     bool            fSnapToDesktopEdges;
     CRect           rcLastWindowPos;
     UINT            nLastWindowType;
